Re-add Provider for global state
