+++ src/services/Rates/index.tsx
@@ -1,0 +1,2 @@
+export { default as useRates, IRatesContext } from './useRates';
+export { RatesContext, RatesProvider } from './RatesProvider';
