+++ src/services/index.ts
@@ -2,6 +2,6 @@ export * from './ApiService';
 export * from './EthService';
 export * from './Store';
 export * from './WalletService';
-export { IRatesContext, default as useRates } from './useRates';
+export * from './Rates';
 export { DevToolsContext, DevToolsProvider, useDevTools } from './DevToolsProvider';
 export { FeatureFlagContext, FeatureFlagProvider, useFeatureFlags } from './FeatureFlagProvider';
